[Dashboard] Add tabindex to panel header tooltip, make it sibling to h2 (#208391)
## Summary
This PR makes tooltip on dashboard panel header keyboard focusable and
moves the tooltip to be a sibling of the `<h2>`.
Closes: #117233

(cherry picked from commit 91120443693e189159681ed1e0c240d1d0886f7e)
